Adds some features in preperation for buiding on civet.
* Adds --purge option for cleaning install directories
* Re-works pages so adding content for each module is simpler
* Adds error counting and non-zero exit code

(refs #6699)

+++ python/MooseDocs/MooseSubApplicationDocGenerator.py
@@ -65,7 +65,7 @@ class MooseSubApplicationDocGenerator(object):
 
         # Purge all files in the install directory
         install_dir = self._config.get('install')
-        if purge:
+        if purge and os.path.exists(install_dir):
             log.info("Purging install directory: {}".format(install_dir))
             shutil.rmtree(install_dir)
